Replace RECOG.TestGroup by new CheckSize helper (#98)
It is not the job of this package to test recog, but the function
`RECOG.TestGroup` is written for that task, and e.g. performs group
recognition multiple times. But we really only want to use recog to compute a
group size in order to test *our* code. So let's do just that, and nothing
more.